control_plane/requirements.txt
anyio==3.6.1
click==8.1.3
-e git+ssh://git@github.com/FlightyAI/flighty.git@7a3aebe16a28567e4acff36d06db090bfcb37011#egg=control_plane&subdirectory=control_plane
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
protobuf==4.21.2
pydantic==1.9.1
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
